Share code between test and analyze command.
